extract useRouteConnect
