+++ extension/src/panel/zodiac-routes/index.ts
@@ -1,4 +1,5 @@
 export { useSelectedRouteId } from './SelectedRouteContext'
+export { useRouteConnect } from './useRouteConnect'
 export { useRouteProvider } from './useRouteProvider'
 export { INITIAL_DEFAULT_ROUTE, useZodiacRoute } from './useZodiacRoute'
 export {
